+++ CHANGELOG.md
@@ -5,10 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [UNRELEASED]
+
+### Added
+
+- ability to change the plugin version if the original version is not found
+- plugins replacing mapping mechanism (TEMPORALLY)
 
 ### Changed
 
+- workspace plugins init mechanism
 - refactored to new router method replace
 
 ## [0.16.1]